Mono-HDR-GS/requirements.txt
torch~=1.13.1+cu117
numpy~=1.24.4
plyfile~=0.8.1
imageio~=2.35.1
pillow~=9.4.0
scipy~=1.10.1
torchvision~=0.14.1+cu117
tqdm~=4.67.1
setuptools~=75.3.0
pyyaml~=6.0.2
pyexr~=0.4.0